fix: move FocusOnHeading to features stories and remove spurious eslint-disable comments
